Rebase storage to SQLite
