--- rest_api/views.py
@@ -1,20 +1,0 @@
-from rest_api import app
-from flask_smorest import Api
-
-from rest_api.resources.user import blp as UserBlueprint
-from rest_api.resources.category import blp as CategoryBlueprint
-from rest_api.resources.record import blp as RecordBlueprint
-
-app.config["PROPAGATE_EXCEPTIONS"] = True
-app.config["API_TITLE"] = "Finance REST API"
-app.config["API_VERSION"] = "v1"
-app.config["OPENAPI_VERSION"] = "3.0.3"
-app.config["OPENAPI_URL_PREFIX"] = "/"
-app.config["OPENAPI_SWAGGER_UO_PATH"] = "/swagger-ui"
-app.config["OPENAPI_SWAGGER_UI_URL"] = "https://cdn.jsdelivr.net/npm/swagger-ui-dist"
-
-api = Api(app)
-
-api.register_blueprint(UserBlueprint)
-api.register_blueprint(CategoryBlueprint)
-api.register_blueprint(RecordBlueprint)
